Last refactor.
