+++ modules/oauth/src/main/java/org/mule/module/oauth2/internal/AbstractGrantType.java
@@ -8,6 +8,8 @@ package org.mule.module.oauth2.internal;
 
 import org.mule.module.http.api.HttpAuthentication;
 import org.mule.module.http.api.requester.proxy.ProxyConfig;
+import org.mule.module.oauth2.internal.tokenmanager.TokenManagerConfig;
+import org.mule.transport.ssl.api.TlsContextFactory;
 
 /**
  * Common interface for all grant types must extend this interface.
@@ -16,6 +18,10 @@ public abstract class AbstractGrantType implements HttpAuthentication, Applicati
 {
 
     protected ProxyConfig proxyConfig;
+    protected TokenManagerConfig tokenManagerConfig;
+    protected TlsContextFactory tlsContextFactory;
+    protected String clientId;
+    protected String clientSecret;
 
     /**
      * @param accessToken an ouath access token
@@ -26,15 +32,43 @@ public abstract class AbstractGrantType implements HttpAuthentication, Applicati
         return "Bearer " + accessToken;
     }
 
-
-    public ProxyConfig getProxyConfig()
-    {
-        return proxyConfig;
-    }
-
     public void setProxyConfig(ProxyConfig proxyConfig)
     {
         this.proxyConfig = proxyConfig;
     }
 
+    public TlsContextFactory getTlsContext()
+    {
+        return tlsContextFactory;
+    }
+
+    public void setTokenManager(TokenManagerConfig tokenManagerConfig)
+    {
+        this.tokenManagerConfig = tokenManagerConfig;
+    }
+
+    public void setTlsContext(TlsContextFactory tlsContextFactory)
+    {
+        this.tlsContextFactory = tlsContextFactory;
+    }
+
+    public void setClientId(final String clientId)
+    {
+        this.clientId = clientId;
+    }
+
+    public void setClientSecret(final String clientSecret)
+    {
+        this.clientSecret = clientSecret;
+    }
+
+    public String getClientSecret()
+    {
+        return clientSecret;
+    }
+
+    public String getClientId()
+    {
+        return clientId;
+    }
 }
